requirements.txt
PyQt6>=6.5
h5py>=3.8
numpy>=1.24
scipy>=1.10
Pillow>=10.0
